VIDEO-9760 Ensure that LocalTrackPublication events are reemitted by the LocalParticipant.
