+++ lib/localparticipant.js
@@ -1,7 +1,14 @@
 'use strict';
 
 const { MediaStreamTrack } = require('./webrtc');
-const { asLocalTrack, asLocalTrackPublication, trackClass } = require('./util');
+
+const {
+  asLocalTrack,
+  asLocalTrackPublication,
+  reemitTrackPublicationEvents,
+  trackClass
+} = require('./util');
+
 const { typeErrors: E, trackPriority } = require('./util/constants');
 const { validateLocalTrack } = require('./util/validate');
 
@@ -118,6 +125,19 @@ class LocalParticipant extends Participant {
       this._addLocalTrack(track, priority);
     }
     return addedTrack;
+  }
+
+  /**
+   * @private
+   * @param {LocalTrackPublication} publication
+   * @returns {?LocalTrackPublication}
+   */
+  _addTrackPublication(publication) {
+    publication = super._addTrackPublication(publication);
+    if (publication) {
+      reemitTrackPublicationEvents(this, publication);
+    }
+    return publication;
   }
 
   /**
