+++ lib/participant.js
@@ -7,7 +7,13 @@ const RemoteDataTrack = require('./media/track/remotedatatrack');
 const RemoteDataTrackPublication = require('./media/track/remotedatatrackpublication');
 const RemoteVideoTrack = require('./media/track/remotevideotrack');
 const RemoteVideoTrackPublication = require('./media/track/remotevideotrackpublication');
-const util = require('./util');
+
+const {
+  reemitTrackPublicationEvents,
+  trackClass,
+  trackPublicationClass,
+  valueToJSON
+} = require('./util');
 
 let nInstances = 0;
 
@@ -226,8 +232,8 @@ class Participant extends EventEmitter {
     tracksByKind.set(key, track);
     reemitTrackEvents(this, track, key);
 
-    log.info(`Added a new ${util.trackClass(track)}:`, key);
-    log.debug(`${util.trackClass(track)}:`, track);
+    log.info(`Added a new ${trackClass(track)}:`, key);
+    log.debug(`${trackClass(track)}:`, track);
 
     return track;
   }
@@ -235,8 +241,8 @@ class Participant extends EventEmitter {
 
   /**
    * @private
-   * @param {RemoteTrackPublication} publication
-   * @returns {?RemoteTrackPublication}
+   * @param {TrackPublication} publication
+   * @returns {?TrackPublication}
    */
   _addTrackPublication(publication) {
     const log = this._log;
@@ -252,8 +258,8 @@ class Participant extends EventEmitter {
     }[publication.kind];
     trackPublicationsByKind.set(publication.trackSid, publication);
 
-    log.info(`Added a new ${util.trackPublicationClass(publication)}:`, publication.trackSid);
-    log.debug(`${util.trackPublicationClass(publication)}:`, publication);
+    log.info(`Added a new ${trackPublicationClass(publication)}:`, publication.trackSid);
+    log.debug(`${trackPublicationClass(publication)}:`, publication);
     return publication;
   }
 
@@ -404,8 +410,8 @@ class Participant extends EventEmitter {
     });
 
     const log = this._log;
-    log.info(`Removed a ${util.trackClass(track)}:`, key);
-    log.debug(`${util.trackClass(track)}:`, track);
+    log.info(`Removed a ${trackClass(track)}:`, key);
+    log.debug(`${trackClass(track)}:`, track);
     return track;
   }
 
@@ -434,13 +440,13 @@ class Participant extends EventEmitter {
     });
 
     const log = this._log;
-    log.info(`Removed a ${util.trackPublicationClass(publication)}:`, publication.trackSid);
-    log.debug(`${util.trackPublicationClass(publication)}:`, publication);
+    log.info(`Removed a ${trackPublicationClass(publication)}:`, publication.trackSid);
+    log.debug(`${trackPublicationClass(publication)}:`, publication);
     return publication;
   }
 
   toJSON() {
-    return util.valueToJSON(this);
+    return valueToJSON(this);
   }
 }
 
@@ -612,34 +618,6 @@ function reemitTrackEvents(participant, track, key) {
   });
 
   participant._trackEventReemitters.set(key, trackEventReemitters);
-}
-
-/**
- * Re-emit {@link TrackPublication} events.
- * @private
- * @param {Participant} participant
- * @param {TrackPublication} publication
- */
-function reemitTrackPublicationEvents(participant, publication) {
-  const publicationEventReemitters = new Map();
-
-  if (participant.state === 'disconnected') {
-    return;
-  }
-
-  participant._getTrackPublicationEvents().forEach(([publicationEvent, participantEvent]) => {
-    publicationEventReemitters.set(publicationEvent, (...args) => {
-      if (publicationEvent === 'trackSwitchedOff') {
-        const [track, switchOffReason] = args;
-        participant.emit(participantEvent, track, publication, switchOffReason);
-      } else {
-        participant.emit(participantEvent, ...args, publication);
-      }
-    });
-    publication.on(publicationEvent, publicationEventReemitters.get(publicationEvent));
-  });
-
-  participant._trackPublicationEventReemitters.set(publication.trackSid, publicationEventReemitters);
 }
 
 module.exports = Participant;
