+++ lib/util/index.js
@@ -851,6 +851,35 @@ function waitForEvent(eventTarget, event) {
   });
 }
 
+
+/**
+ * Re-emit {@link TrackPublication} events.
+ * @private
+ * @param {Participant} participant
+ * @param {TrackPublication} publication
+ */
+function reemitTrackPublicationEvents(participant, publication) {
+  const publicationEventReemitters = new Map();
+
+  if (participant.state === 'disconnected') {
+    return;
+  }
+
+  participant._getTrackPublicationEvents().forEach(([publicationEvent, participantEvent]) => {
+    publicationEventReemitters.set(publicationEvent, (...args) => {
+      if (publicationEvent === 'trackSwitchedOff') {
+        const [track, switchOffReason] = args;
+        participant.emit(participantEvent, track, publication, switchOffReason);
+      } else {
+        participant.emit(participantEvent, ...args, publication);
+      }
+    });
+    publication.on(publicationEvent, publicationEventReemitters.get(publicationEvent));
+  });
+
+  participant._trackPublicationEventReemitters.set(publication.trackSid, publicationEventReemitters);
+}
+
 exports.constants = constants;
 exports.createBandwidthProfilePayload = createBandwidthProfilePayload;
 exports.createMediaSignalingPayload = createMediaSignalingPayload;
@@ -887,3 +916,4 @@ exports.isChromeScreenShareTrack = isChromeScreenShareTrack;
 exports.isUserMediaTrack = isUserMediaTrack;
 exports.waitForSometime = waitForSometime;
 exports.waitForEvent = waitForEvent;
+exports.reemitTrackPublicationEvents = reemitTrackPublicationEvents;
